refactor(#3037): relocate escalation settings DTOs to config layer (backflow 39→35)
Move `EscalationScheduleSettings`, `EscalationSettings`, and
`EscalationSettingsResponse` (plus the `DEFAULT_ESCALATION_PM_HOURS` /
`DEFAULT_ESCALATION_TIMEZONE` constants) out of the server route module
`src/server/routes/escalation.rs` into `src/config.rs`, alongside the
related `EscalationConfig` / `EscalationMode` domain types.

These are pure serde data types (the materialized wire/runtime shape of
escalation settings) that the service layer consumes when calling the
internal settings API. Housing them in the route module forced
`services/discord/commands/text_commands.rs` and
`services/discord/internal_api.rs` to reach back into `crate::server`
(service→server backflow, #3037).

config is a leaf domain module, so depending on it from both server and
services introduces no cycle. Behavior is unchanged — pure relocation +
call-site repointing. escalation.rs and internal_api.rs reuse the moved
types via `use`.

Backflow (live audit, comment-stripped): 39→35; the `text_commands.rs`
cluster (4 findings) is fully eliminated. Baseline regenerated from the
live measurement (monotonic decrease; gate satisfied).

Gates: cargo fmt --check clean; cargo check --lib clean (0 new warnings);
cargo test --lib escalation/config passing; audit --check exit 0;
generate_inventory_docs + check_agent_maintenance_docs freshness passed
(change-surfaces freeze entries synced to generated truth);
ci-script-checks.sh exit 0.
